Start moving stuff into functions folder.
