+++ csharp/src/sdk/Functions/HyparConfig.cs
@@ -1,7 +1,7 @@
 using Newtonsoft.Json;
 using System.Collections.Generic;
 
-namespace Hypar.Configuration
+namespace Hypar.Functions
 {
     /// <summary>
     /// A container for Hypar configuration information.
@@ -13,12 +13,6 @@ namespace Hypar.Configuration
         /// </summary>
         [JsonProperty("description")]
         public string Description{get;set;}
-
-        /// <summary>
-        /// The fully-qualified name of the function.
-        /// </summary>
-        [JsonProperty("function")]
-        public string Function{get;set;}
 
         /// <summary>
         /// The unique identifier of the function.
@@ -33,16 +27,10 @@ namespace Hypar.Configuration
         public string Name{get;set;}
 
         /// <summary>
-        /// The runtime used to execute the function.
-        /// </summary>
-        [JsonProperty("runtime")]
-        public string Runtime{get;set;}
-
-        /// <summary>
         /// A map of input parameter data for the function.
         /// </summary>
         [JsonProperty("parameters")]
-        public Dictionary<string,ParameterData> Parameters{get;set;}
+        public Dictionary<string,ParameterBase> Parameters{get;set;}
 
         /// <summary>
         /// An optional git repository that stores your function.
@@ -81,7 +69,7 @@ namespace Hypar.Configuration
         /// </summary>
         public HyparConfig()
         {
-            this.Parameters = new Dictionary<string, ParameterData>();
+            this.Parameters = new Dictionary<string, ParameterBase>();
             this.Returns = new Dictionary<string, ReturnData>();
         }
     }
