+++ csharp/src/sdk/Functions/LocationParameter.cs
@@ -1,0 +1,14 @@
+namespace Hypar.Functions
+{
+    /// <summary>
+    /// A location parameter.
+    /// </summary>
+    public class LocationParameter: ParameterBase
+    {
+        /// <summary>
+        /// Construct a location parameter.
+        /// </summary>
+        /// <param name="description"></param>
+        public LocationParameter(string description):base(description, ParameterType.Location){}
+    }
+}
