+++ csharp/test/Hypar.SDK.Tests/ExecutionTests.cs
@@ -8,12 +8,27 @@ using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
 using System.IO;
+using Hypar.Functions;
 using System.Linq;
 
 using Xunit;
 
 namespace Hypar.Tests
 {
+    internal class TestFunction : IHyparFunction
+    {
+        public Model Execute(Model model, Dictionary<string, object> parameters, Dictionary<string, object> returns)
+        {
+            var profile = (Hypar.Geometry.Polygon)parameters["location"];
+            var origin = (Position)parameters["origin"];
+
+            var mass = new Mass(profile, 0.0, 5.0);
+            model.AddElement(mass);
+            model.Origin = origin;
+            return model;
+        }
+    }
+
     public class ExecutionTests
     {
         [Fact]
@@ -29,11 +44,11 @@ namespace Hypar.Tests
             var plines = outline.ToPolygons();
             var transformed = plines.Select(pline=>new Hypar.Geometry.Polygon(pline.Vertices.Select(v=>new Vector3(v.X - origin.X, v.Y - origin.Y, v.Z)).ToArray()).Reversed()).ToArray();
 
-            var mass = new Mass(transformed[0], 0.0, 5.0);
-            var model = new Model();
-            model.Origin = outline.Coordinates[0][0];
+            // Compare the request to the 
+            // Execute the function
+            var func = new TestFunction();
+            var model = func.Execute(new Model(), )
 
-            model.AddElement(mass);
             model.SaveGlb("siteMass.glb");
             
             var json = JsonConvert.SerializeObject(model);
